+++ benchmark.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+for engine in fast precise; do
+    for drift in 1 16 32 64 128; do
+        for window in 1 16 32 64 128; do
+            base="o-${engine}-d${drift}-w${window}"
+            echo ${base}
+            time python -m vidkompy -b tests/bg.mp4 -f tests/fg.mp4 -e ${engine} -d ${drift} -w ${window} -o tests/${base}.mp4 --verbose >tests/${base}.txt 2>tests/${base}.err.txt
+        done
+    done
+done